Add files via upload
I've got the Racetrack and my Traction Engine in this ZIP as well as the texture for the road. I did some testing with my traction engine beforehand and it seems the parts move okay when scaled so you can scale my truck/road to your size or vice versa if you want and there shouldn't be any problems.